requirements.txt
pytest~=7.1.2
easyocr~=1.5.0
PyAutoGUI~=0.9.53
Pillow~=9.2.0